+++ Src/MoneyFox.Uwp/Views/Accounts/AddAccountDialog.xaml
@@ -1,11 +1,10 @@
 ﻿<ContentDialog
-    x:Class="MoneyFox.Uwp.Views.AddAccountDialog"
+    x:Class="MoneyFox.Uwp.Views.Accounts.AddAccountDialog"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:resources="using:MoneyFox.Application.Resources"
-    xmlns:userControls="using:MoneyFox.Uwp.Views.UserControls"
+    xmlns:resources="using:MoneyFox.Application.Resources" xmlns:userControls="using:MoneyFox.Uwp.Views.Accounts"
     Title="{Binding Title}"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
     PrimaryButtonCommand="{Binding SaveCommand}"
