+++ Src/MoneyFox.Uwp/Views/Accounts/ModifyAccountUserControl.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="MoneyFox.Uwp.Views.UserControls.ModifyAccountUserControl"
+    x:Class="MoneyFox.Uwp.Views.Accounts.ModifyAccountUserControl"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:core="using:Microsoft.Xaml.Interactions.Core"
