+++ Src/MoneyFox.Uwp/Views/Categories/AddCategoryDialog.xaml
@@ -1,11 +1,11 @@
 ﻿<ContentDialog
-    x:Class="MoneyFox.Uwp.Views.AddCategoryDialog"
+    x:Class="MoneyFox.Uwp.Views.Categories.AddCategoryDialog"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:resources="using:MoneyFox.Application.Resources"
-    xmlns:userControls="using:MoneyFox.Uwp.Views.UserControls"
+    xmlns:userControls="using:MoneyFox.Uwp.Views.Categories"
     Title="{Binding Title}"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
     PrimaryButtonCommand="{Binding SaveCommand}"
