+++ Src/MoneyFox.Uwp/Views/Categories/CategoryListControl.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="MoneyFox.Uwp.Views.UserControls.CategoryListControl"
+    x:Class="MoneyFox.Uwp.Views.Categories.CategoryListControl"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:controls="using:Microsoft.Toolkit.Uwp.UI.Controls"
