+++ Src/MoneyFox.Uwp/Views/Categories/EditCategoryDialog.xaml.cs
@@ -2,7 +2,7 @@
 using Windows.UI.Xaml.Controls;
 
 #nullable enable
-namespace MoneyFox.Uwp.Views
+namespace MoneyFox.Uwp.Views.Categories
 {
     public sealed partial class EditCategoryDialog : ContentDialog
     {
