+++ Src/MoneyFox.Uwp/Views/Categories/ModifyCategoryUserControl.xaml.cs
@@ -1,5 +1,5 @@
 ﻿#nullable enable
-namespace MoneyFox.Uwp.Views.UserControls
+namespace MoneyFox.Uwp.Views.Categories
 {
     public sealed partial class ModifyCategoryUserControl
     {
